Do not store sertype pointer in topic definition
Remove the pointer to the sertype from struct topic_definition.
This avoids a locking order issue (locking the proxy participant
while already holding tl_admin_lock) when updating sertype pointers
in topic definitions for proxy topics in case a new type is
registered.

The sertype pointer was used in the find remote topic implementation,
but this can easily be worked around by a lookup of the sertype via
the tl_admin using the type identifier stored in the topic definition.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>