Replace type_id by tl_meta pointer
Replace the type identifier in (proxy) endpoints and topic definitions
by a pointer to the type-lookup meta data. This change is to prepare
for the implementation of the XTypes type system, in which a type
identifier can't be used as an identifier in certain cases, because
a minimal and complete version of the type identifier exists.

Signed-off-by: Dennis Potman <dennis.potman@adlinktech.com>